Refactor: 전표 테스트
수금 및 세금계산서 controllerTest refactor
